Okabe

+++ Assets/Player/PlayerStatusData.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "Data/Create StatusData")]
 public class StatusData : ScriptableObject
 {
-    public float _hp;//HP
+    public int _hp;//HP
     public float _moveSpeed; //移動速度
 }
